==> 04) dependencies_adjectives_process.ipynb <==
import pandas as pd
from tqdm import tqdm

tqdm.pandas()

Shortcuts:
- Allready separated ngrams further down
- Year/depency/adjective summarized

df = pd.read_csv('./data/ngram_dependencies_adjectives.csv')
print(df.head())
print(df.info())

df['ngram'] = df['ngram'].str.replace('_adj', '', n=2)
df.head()

df[['dependency', 'adjective']] = df['ngram'].str.split('=>', expand=True)
df.head()

df_no_ngram = df.drop('ngram', 1)

df_no_ngram.to_csv('./data/ngram_dependencies_adjectives_separated.csv', index=False)

df_no_ngram = pd.read_csv('./data/ngram_dependencies_adjectives_separated.csv')
df_no_ngram.head()

TOP_N = 1000

adjective_occurences = df_no_ngram\
    .groupby('adjective')\
    .sum()\
    .sort_values('count', ascending=False)\
    .reset_index()

tops = adjective_occurences['adjective'][:TOP_N]
tops_set = set(tops)

df_tops = df_no_ngram[df_no_ngram['adjective'].progress_apply(lambda x: x in tops_set)]

print(f'Size before: {df_no_ngram.shape}')
print(f'Size after: {df_tops.shape}')

df_summarized = df_tops.groupby(['dependency', 'year', 'adjective']).sum().reset_index()

df_summarized = df_summarized.reset_index().rename(columns={'year': 'calendar_year'})

df_summarized.to_csv("./data/ngram_dependencies_adjectives_separated_summarized.csv", index=False)

df_summarized = pd.read_csv("./data/ngram_dependencies_adjectives_separated_summarized.csv")

df_wide = pd.pivot(df_summarized, index=['dependency', 'calendar_year'], columns='adjective', values='count').reset_index()

df_wide_filled = df_wide.fillna(0)

print(df_wide_filled.info())
df_wide_filled.head()

df_wide_filled_indices = df_wide_filled.iloc[:, :2]
df_wide_filled_features = df_wide_filled.iloc[:, 2:]
row_mean = df_wide_filled_features.mean(axis=1)
row_std = df_wide_filled_features.std(axis=1)

df_wide_filled_features_normalized = df_wide_filled_features.subtract(row_mean, axis=0).divide(row_std, axis=0)
df_wide_filled_normalized = pd.concat([df_wide_filled_indices, df_wide_filled_features_normalized], axis=1)

print(df_wide_filled_normalized.info())
df_wide_filled_normalized.head()

df_wide_filled.to_csv("./data/ngram_dependencies_adjectives_wide.csv", index=False)


df_wide_filled_normalized.to_csv("./data/ngram_dependencies_adjectives_wide_normalized.csv", index=False)
