==> 05) Clustering xs by dependencies.ipynb <==
import pandas as pd
import tensorflow.keras as k
from sklearn.preprocessing import StandardScaler

adjective_data = pd.read_csv('./data/ngram_dependencies_adjectives_wide_normalized.csv')
print(adjective_data.columns[:10])

X = adjective_data.iloc[:, 2:]
X = StandardScaler().fit_transform(X)


def get_dense_relu(x): return k.layers.Dense(x, activation="relu")  #, kernel_regularizer="l2")

n_vars = X.shape[1]

model = k.Sequential([
    k.layers.Dense(500, input_shape=[n_vars], activation="relu"),
    get_dense_relu(200),
    get_dense_relu(50),
    get_dense_relu(10),
    k.layers.Dense(2, name="encoded", activation="sigmoid"),
    get_dense_relu(10),
    get_dense_relu(50),
    get_dense_relu(200),
    k.layers.Dense(n_vars)
])
model.compile(optimizer="adam", loss="mse")
# model.summary()

callback_save_checkpoint = k.callbacks.ModelCheckpoint(
    './data/model_checkpoints/Epoch{epoch:02d}-Loss{loss:.2f}.hdf5',
    monitor='loss',
    verbose=0,
    save_best_only=False,
    save_weights_only=False,
    mode='auto',
    save_freq='epoch'
)

early_stopping = k.callbacks.EarlyStopping(
    "loss",
    min_delta=0.0001,
    patience=3
)

history = model.fit(X, X, batch_size=100, epochs=1000, shuffle=True, callbacks=[early_stopping])  # , callbacks=[callback_save_checkpoint]

k.models.save_model(model, "./data/model_autoencoder.hdf5")
